Store first implementation finished; tests added

+++ internal/app/models/models.go
@@ -6,8 +6,8 @@ import (
 
 // Account holds id and amount of money
 type Account struct {
-	AccountID int64 `json:"account_id"`
-	Balance   int64 `json:"balance"`
+	AccountID int64   `json:"account_id"`
+	Balance   float64 `json:"balance"`
 }
 
 // Transfer holds data needed to perform money transfer
@@ -16,12 +16,12 @@ type Transfer struct {
 	Timestamp     time.Time `json:"timestamp"`
 	FromAccountID int64     `json:"from_account_id"`
 	ToAccountID   int64     `json:"to_account_id"`
-	Amount        int64     `json:"amount"`
+	Amount        float64   `json:"amount"`
 }
 
 // TransferResult hodls data about performed transfer and the new balances
 type TransferResult struct {
 	Transfer             Transfer `json:"transfer"`
-	FromAccountIDBalance int64    `json:"from_account_id_balance"`
-	ToAccountIDBalance   int64    `json:"to_account_id_balance"`
+	FromAccountIDBalance float64  `json:"from_account_id_balance"`
+	ToAccountIDBalance   float64  `json:"to_account_id_balance"`
 }
